# backend/openapi.yaml
openapi: 3.0.0
info:
  title: 'Chiorino ERP API'
  version: '1'
paths:
...
  /order-split:
    get:
      tags:
        - order-split
      summary: 'List of lot to split and inserted cuts'
      operationId: 1c501228391bd9f36dabefbc3f5b73f8
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - idStock
              properties:
                idStock:
                  type: integer
              type: object
            example:
              idStock: 41224
      responses:
        '200':
          description: OK
        '500':
          description: 'Internal Error'
        '401':
          description: Unauthorized
    post:
      tags:
        - order-split
      summary: 'Add split to lot'
      operationId: adf4826be976b48efa8e3c23c9d52f82
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - idStock
                - qty
                - idWarehouseLocation
              properties:
                idStock:
                  type: integer
                qty:
                  type: int|float
                ordRef:
                  type: string|null
                idWarehouseLocation:
                  type: integer
              type: object
            example:
              idStock: 183151
              ordRef: ''
              qty: 1
              idWarehouseLocation: 1954
      responses:
        '200':
          description: OK
        '500':
          description: 'Internal Error'
        '401':
          description: Unauthorized
  '/order-split/{id}':
    delete:
      tags:
        - order-split
      summary: 'Delete split row'
      operationId: 31055eefb6981171668f9631893c9b11
      parameters:
        -
          $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
        '403':
          description: Unauthorized
        '404':
          description: 'Not Found'
        '500':
          description: 'Internal Error'
  /order-split/confirm:
    post:
      tags:
        - order-split
      summary: 'Confirm split order'
      operationId: 8692e9fc37601db3780b3d6904dcd43c
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - idStock
              properties:
                idStock:
                  type: integer
              type: object
            example:
              idStock: 183153
      responses:
        '200':
          description: OK
        '500':
          description: 'Internal Error'
        '401':
          description: Unauthorized
...
  /sales-total-discount-matrix:
    get:
      tags:
        - 'Sales Total Discount Matrix'
      summary: 'Get sales total discount matrix'
      operationId: 4300a26d15ba15f2693ee95bd3785050
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SaleTotalDiscountMatrixResource' } }
                type: object
    post:
      tags:
        - 'Sales Total Discount Matrix'
      summary: 'Create sales total discount matrix'
      operationId: 2c21c33586b76a7d17518a7e91bef174
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - priority
                - description
              properties:
                priority:
                  type: integer
                description:
                  type: string
              type: object
            example:
              priority: 1
              description: 'Matrice 1'
              currency_id: EUR
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SaleTotalDiscountMatrixResource' } }
                type: object
...
  '/sales-total-discount-matrix/{id}/toggle':
    put:
      tags:
        - 'Sales Total Discount Matrix'
      summary: 'Enable/disable sales total discount matrix'
      operationId: 0eb72b48f1d761b3cd30d8e73fd69757
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - disable
              properties:
                disable:
                  type: boolean
              type: object
            example:
              disable: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SaleTotalDiscountMatrixResource' } }
                type: object
...